--- packages/styled-system/src/create-theme-vars/transform-template.ts
@@ -1,28 +1,0 @@
-/**
- * The CSS transform order following the upcoming spec from CSSWG
- * translate => rotate => scale => skew
- * @see https://drafts.csswg.org/css-transforms-2/#ctm
- * @see https://www.stefanjudis.com/blog/order-in-css-transformation-transform-functions-vs-individual-transforms/
- */
-const transformTemplate = [
-  "rotate(var(--chakra-rotate, 0))",
-  "scaleX(var(--chakra-scale-x, 1))",
-  "scaleY(var(--chakra-scale-y, 1))",
-  "skewX(var(--chakra-skew-x, 0))",
-  "skewY(var(--chakra-skew-y, 0))",
-]
-
-export function getTransformTemplate() {
-  return [
-    "translateX(var(--chakra-translate-x, 0))",
-    "translateY(var(--chakra-translate-y, 0))",
-    ...transformTemplate,
-  ].join(" ")
-}
-
-export function getTransformGpuTemplate() {
-  return [
-    "translate3d(var(--chakra-translate-x, 0), var(--chakra-translate-y, 0), 0)",
-    ...transformTemplate,
-  ].join(" ")
-}
