+++ packages/styled-system/src/index.ts
@@ -1,6 +1,6 @@
 export * from "./config"
 export * from "./css"
-export * from "./types"
+export * from "./system.types"
 export * from "./theming.types"
 export * from "./system"
 export * from "./create-theme-vars"
