+++ packages/styled-system/src/system.types.ts
@@ -4,16 +4,16 @@ import type {
   BackgroundProps,
   BorderProps,
   ColorProps,
+  EffectProps,
+  FilterProps,
   FlexboxProps,
   GridProps,
-  FilterProps,
+  InteractivityProps,
   LayoutProps,
   ListProps,
   OtherProps,
   PositionProps,
   RingProps,
-  InteractivityProps,
-  EffectProps,
   SpaceProps,
   TextDecorationProps,
   TransformProps,
